requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-multiselectfield==0.1.12
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
